+++ cognite/neat/rules/importer/__init__.py
@@ -3,7 +3,7 @@ from ._dict2rules import ArbitraryDictImporter
 from ._dms2rules import DMSImporter
 from ._graph2rules import GraphImporter
 from ._json2rules import ArbitraryJSONImporter
-from ._owl._owl2rules import OWLImporter
+from ._owl2rules import OWLImporter
 from ._spreadsheet2rules import ExcelImporter, GoogleSheetImporter
 from ._xml2rules import XMLImporter
 from ._yaml2rules import ArbitraryYAMLImporter
