+++ cognite/neat/rules/importer/_base.py
@@ -19,16 +19,14 @@ class BaseImporter(ABC):
         ...
 
     @abstractmethod
-    def to_tables(self, use_default_values: bool | None = None) -> dict[str, pd.DataFrame]:
+    def to_tables(self) -> dict[str, pd.DataFrame]:
         """Creates raw tables from the data."""
         raise NotImplementedError
 
-    def to_raw_rules(self, use_default_values: bool | None = None) -> RawRules:
+    def to_raw_rules(self) -> RawRules:
         """Creates `RawRules` object from the data."""
-        if use_default_values:
-            tables = self.to_tables(use_default_values=use_default_values)
-        else:
-            tables = self.to_tables()
+
+        tables = self.to_tables()
 
         return RawRules.from_tables(tables=tables, importer_type=self.__class__.__name__)
 
@@ -37,7 +35,6 @@ class BaseImporter(ABC):
         return_report: bool = False,
         skip_validation: bool = False,
         validators_to_skip: list[str] | None = None,
-        use_default_values: bool | None = None,
     ) -> tuple[Rules | None, list[ErrorDetails] | None, list | None] | Rules:
         """
         Creates `Rules` object from the data.
@@ -53,14 +50,12 @@ class BaseImporter(ABC):
             and optional list of errors and warnings if
             `return_report` is set to True.
 
-        !!! Note
+        !!! Note "Skip Validation
             `skip_validation` flag should be only used for purpose when `Rules` object
             is exported to an Excel file. Do not use this flag for any other purpose!
         """
-        if use_default_values:
-            raw_rules = self.to_raw_rules(use_default_values=use_default_values)
-        else:
-            raw_rules = self.to_raw_rules()
+
+        raw_rules = self.to_raw_rules()
 
         return raw_rules.to_rules(return_report, skip_validation, validators_to_skip)
