+++ cognite/neat/rules/importer/_owl2rules/__init__.py
@@ -1,0 +1,3 @@
+from ._owl2rules import OWLImporter
+
+__all__ = ["OWLImporter"]
